refactor: Update PostgreSQL image version to 16 in docker-compose.yml
Update the PostgreSQL image version to 16 in the docker-compose.yml file. This ensures that the application is using the latest version of PostgreSQL.

refactor: Update SQL queries to include position field for product images

Update the SQL queries in the admin_requests.sql and users.sql files to include the position field for product images. This allows for better organization and ordering of product images.

refactor: Add GetAdminRequestByUserID method to AdminRequestRepository

Add the GetAdminRequestByUserID method to the AdminRequestRepository interface and implement it in the admin_request_repository.go file. This method retrieves an admin request by user ID and status.

refactor: Add GetUserProfileByID and UpdateUserInfo methods to UserRepository

Add the GetUserProfileByID and UpdateUserInfo methods to the UserRepository interface and implement them in the user_repository.go file. These methods retrieve a user's profile by ID and update a user's information, respectively.